week2-multi-agent-poc/requirements.txt
groq
boto3
python-dotenv